+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [2.3.0] - 2021-XX-XX
 - Deprecation warnings added to:
+  - Instrument class (old meta labels)
    - Custom class
 - Documentation
    - Updated docstrings with deprecation notes